사장님 커뮤니티 이후성 컨트롤러 정리, 상세페이지 UI정리
