+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/boss/ChartVO.java
@@ -7,7 +7,8 @@ public class ChartVO {
 	private String date;
     private int total_amount;
     private String order_id;
-	
+    
+    // getters and setters
 	public String getOrder_id() {
 		return order_id;
 	}
